pyproject.toml
[tool.poetry]
name = "main"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
flask = {extras = ["async"], version = "^3.0.3"}
pyinstaller = "^6.4.0"
pytz = "^2024.1"
loguru = "^0.7.2"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
tomli = "^2.0.1"
aiohttp = "^3.9.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
